+++ src/Controller/ProductManager/ProductManagerTrashCanController.php
@@ -12,6 +12,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\Security\Core\Security;
 
 
 class ProductManagerTrashCanController extends AbstractController{
@@ -26,19 +27,37 @@ class ProductManagerTrashCanController extends AbstractController{
      */
     private $em;
 
-    public function __construct(TrashCanRepository $trashCanRep, EntityManagerInterface $em)
+    /**
+     * @var Security
+     */
+    private $security;
+
+    /**
+     * ProductManagerTrashCanController constructor.
+     * @param TrashCanRepository $trashCanRep
+     * @param EntityManagerInterface $em
+     * @param Security $security
+     */
+    public function __construct(TrashCanRepository $trashCanRep, EntityManagerInterface $em, Security $security)
     {
         $this->trashCanRep = $trashCanRep;
         $this->em = $em;
+        $this->security = $security;
     }
 
     /**
      * @Route("/productmanager/trashCan", name="productmanager.trashCan.index")
      *
+     * @param PaginatorInterface $paginatorInterface
+     * @param Request $request
      * @return Response
      */
     public function index(PaginatorInterface $paginatorInterface, Request $request)
     {
+        if (!$this->security->getUser()->getActivate() && !$this->get('security.authorization_checker')->isGranted('ROLE_ADMIN')) {
+            throw $this->createAccessDeniedException('Accès refusé, compte désactivé');
+        }
+
         $trashCans = $paginatorInterface->paginate(
             $this->trashCanRep->findAllOrderQuery(),
             $request->query->getInt('page', 1),
@@ -59,6 +78,10 @@ class ProductManagerTrashCanController extends AbstractController{
      */
     public function new(Request $request)
     {
+        if (!$this->security->getUser()->getActivate() && !$this->get('security.authorization_checker')->isGranted('ROLE_ADMIN')) {
+            throw $this->createAccessDeniedException('Accès refusé, compte désactivé');
+        }
+
         $trashCan = new TrashCan();
 
         $form = $this->createForm(TrashCanType::class, $trashCan);
@@ -86,6 +109,9 @@ class ProductManagerTrashCanController extends AbstractController{
      */
     public function edit(TrashCan $trashCan, Request $request)
     {
+        if (!$this->security->getUser()->getActivate() && !$this->get('security.authorization_checker')->isGranted('ROLE_ADMIN')) {
+            throw $this->createAccessDeniedException('Accès refusé, compte désactivé');
+        }
 
         $form = $this->createForm(TrashCanType::class, $trashCan);
         $form->handleRequest($request);
@@ -117,6 +143,9 @@ class ProductManagerTrashCanController extends AbstractController{
      */
     public function delete(TrashCan $trashCan, Request $request)
     {
+        if (!$this->security->getUser()->getActivate() && !$this->get('security.authorization_checker')->isGranted('ROLE_ADMIN')) {
+            throw $this->createAccessDeniedException('Accès refusé, compte désactivé');
+        }
 
         if ($this->isCsrfTokenValid('delete' . $trashCan->getId(), $request->get('_token'))) {
             $this->em->remove($trashCan);
@@ -129,10 +158,14 @@ class ProductManagerTrashCanController extends AbstractController{
     /**
      * @Route("/productmanager/cancel/trashCan", name="productmanager.trashCan.cancel")
      *
+     * @param Request $request
      * @return RedirectResponse
      */
     public function cancel(Request $request)
     {
+        if (!$this->security->getUser()->getActivate() && !$this->get('security.authorization_checker')->isGranted('ROLE_ADMIN')) {
+            throw $this->createAccessDeniedException('Accès refusé, compte désactivé');
+        }
 
         $this->addFlash('warning', 'Poubelle non enregistré');
